Added some new methods and created option to store the ssh key used t…

+++ src/git-key.js
@@ -1,10 +1,12 @@
 const fs = require('fs');
 const path = require('path');
+const { v4: uuidv4 } = require('uuid');
 
 class GitKey{
 
-  constructor(keyPath){
+  constructor(keyPath, saveCreds){
     this.keyPath = keyPath;
+    this.saveCreds = saveCreds;
   }
 
   /**
@@ -12,14 +14,14 @@ class GitKey{
    * @param {*} keyParam 
    * @returns {Promise<GitKey>}
    */
-  static async from(keyParam){
+  static async from(keyParam, saveCreds){
     if (!keyParam){
       throw "SSH key must be specified";
     }
     const key = keyParam.replace(/\\n/g,'\n');
     
     // Write private key to file
-    const keyFileName = `git-key-${new Date().getTime()}.pem`;
+    const keyFileName = `git-key-${uuidv4()}.pem`;
     const keyPath = path.join(__dirname, keyFileName);
     return new Promise((resolve,reject)=>{
       fs.writeFile(keyPath, key,(err)=>{
@@ -28,14 +30,39 @@ class GitKey{
         // Set key file permissions
         fs.chmod(keyPath, '0400', (error) => {
           if(error) return reject(error);
-          const gitKey = new GitKey(keyPath)
+          const gitKey = new GitKey(keyPath, saveCreds)
           resolve(gitKey);
         });
       })
     });
   }
 
+  
+  /**
+   * 
+   * @param {string} path 
+   * @returns {Promise<GitKey | null>}
+   */
+   static async fromRepoFolder(path){
+    const {execGitCommand} = require('./helpers');
+    if (!path){
+      throw "Must provide repository path";
+    }
+    try {
+      const sshCommand = await execGitCommand(["config --get core.sshCommand"], path);
+      const keyPath = sshCommand.match(/-i ([^\n\r]+)/)[1].replace(/\\\\/g, "\\");
+      if (!keyPath) return null;
+      return new GitKey(keyPath);
+    }
+    catch (err) {
+      console.error(err);
+      throw err;
+      return null;
+    }
+  }
+
   async dispose(){
+    if (this.saveCreds) return;
     const self = this;
     return new Promise((resolve,reject)=>{
       fs.unlink(self.keyPath,(err)=>{
